requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.10.4
requests
SQLAlchemy==2.0.36
